fixed homepage searchbar/date/time/category filter
