+++ client/src/components/homepage/homepage.jsx
@@ -23,41 +23,21 @@ import "../../../../node_modules/react-datepicker/dist/react-datepicker.css";
 import TimeRangePicker from "@wojtekmaj/react-timerange-picker/dist/entry.nostyle";
 import axios from "axios";
 import Event from "./event.jsx";
+import moment from "moment";
 import { MdSettingsBackupRestore } from "react-icons/md";
 
-const HomePage = ({ searchEvent }) => {
-  const [isLoading, setIsLoading] = useState(true);
-  const [reset, setReset] = useState(false);
+const HomePage = ({ reset, setReset, events, setEvents, searchEvent, categoriesList }) => {
 
   const categories = useRef(null);
   const slideLeft = useRef(null);
   const slideRight = useRef(null);
 
-  const [events, setEvents] = useState([]);
-  const [categoriesList, setCategoriesList] = useState([]);
 
   const [startDate, setStartDate] = useState(new Date());
   const [label, setLabel] = useState("");
   const [initial, setInitial] = useState(true);
   const [duration, setDuration] = useState("");
 
-  useEffect(() => {
-    const getEvents = axios.get("/api/events", { params: { limit: 10, page: 0 } })
-      .then((response) => {
-        setEvents(response.data);
-      });
-    const getAllCategories = axios.get("/api/categories").then((response) => {
-      setCategoriesList(response.data);
-    });
-    const promises = [getEvents, getAllCategories];
-    Promise.all(promises)
-      .then(() => {
-        setIsLoading(false);
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  }, [reset]);
 
   useEffect(() => {
     if (searchEvent.length > 0) {
@@ -104,11 +84,11 @@ const HomePage = ({ searchEvent }) => {
 
   const searchEventsTime = () => {
     const newDuration = duration.substring(0, duration.length - 5);
-    const newDate = startDate.toISOString().slice(0, 10);
+    const newDate = moment(startDate).format().slice(0, 10);
     const newEvents = events.filter(
       (event, idx) =>
         event.event_length_minutes == newDuration &&
-        event.date.slice(0, 10) === newDate
+        event.date.slice(0, 10) == newDate
     );
     setEvents(newEvents);
   };
